Add a public `teardown` method to `Trainer`
