+++ src/lightning/pytorch/loops/evaluation_loop.py
@@ -304,11 +304,6 @@ class _EvaluationLoop(_Loop):
         if self._data_fetcher is not None:
             self._data_fetcher.teardown()
             self._data_fetcher = None
-
-        if self._data_source is not None:
-            self._data_source.teardown()
-        self._combined_loader = None
-
         self._results.cpu()
 
     def _on_evaluation_start(self, *args: Any, **kwargs: Any) -> None:
